Factor out IconDataRow from CurrencyRow
We need to begin displaying other non-wallet data in the same visual format. Split up CurrencyRow to support this.

Style names were retained to improve reviewability.

Integrate IconDataRow into CurrencyRow
